Redirect signed-in users away from login/register/reset pages
Login honors next=, register and reset send them to / and
/account/settings respectively, matching how login already redirects
post-auth.
