+++ src/app/account/reset/page.tsx
@@ -1,48 +1,19 @@
-'use client'
+import { redirect } from 'next/navigation'
 
-import { useState } from 'react'
+import { createClient } from '@/utils/supabase/server'
 
-import { reset } from './actions'
+import { ResetForm } from './resetForm'
 
-const ResetPage = () => {
-  const [disabled, setDisabled] = useState(false)
-  const [emailSent, setEmailSent] = useState(false)
-
-  const resetAndReturn = async (formData: FormData) => {
-    await reset(formData)
-    setEmailSent(true)
+const ResetPage = async () => {
+  const supabase = await createClient()
+  const {
+    data: { user },
+  } = await supabase.auth.getUser()
+  if (user) {
+    redirect('/account/settings')
   }
 
-  const handleEmailChange = () => {
-    setDisabled(false)
-  }
-
-  return (
-    <form
-      onSubmit={() => {
-        setDisabled(true)
-        setEmailSent(false)
-      }}
-    >
-      <h1>Reset Password</h1>
-      <p>Forgot your password? Let&apos;s confirm your email to reset it.</p>
-      <label htmlFor="email">Email:</label>
-      <br />
-      <input id="email" name="email" type="email" required onChange={handleEmailChange} />
-      <br />
-      <button formAction={resetAndReturn} disabled={disabled}>
-        Send Email
-      </button>
-      {emailSent && (
-        <>
-          <svg height="10" width="20">
-            <circle cx="10" cy="5" r="5" fill="#00AF00" />
-          </svg>
-          Check your email for a link.
-        </>
-      )}
-    </form>
-  )
+  return <ResetForm />
 }
 
 export default ResetPage
